코드 리팩토링
- CoroutineScope 반영하여 기본적인 Thread 관리
- 패키지 경로 변경
- ActionBar 삭제
