+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
         <activity android:name=".TakingPictureActivity"/>
         <activity android:name=".PixelizedImageActivity" />
         <activity android:name=".ImageProcessActivity" />
-        <activity android:name=".PixelActivity" />
+        <activity android:name=".presentation.pixels.PixelActivity" />
     </application>
 
 </manifest>
