+++ app/src/main/res/layout/activity_pixel.xml
@@ -5,7 +5,7 @@
     android:id="@+id/canvas_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PixelActivity">
+    tools:context=".presentation.pixels.PixelActivity">
 
     <com.bizzarestudy.imagecolorpicker.presentation.pixels.PixelCanvas
         android:id="@+id/pixel_canvas"
